% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/application_metadata_applicants.R
\docType{data}
\name{application_metadata_applicants}
\alias{application_metadata_applicants}
\title{application_metadata_applicants: ETH Board Open Research Data (ORD) Program Project Metadata and Report Data}
\format{
A data frame with 113 rows and 7 variables:
\describe{
\item{project_id}{Unique identifier for each project within the ETH Open Research Data dataset. (Examples: CONTRIBUTE/SOLARFUELSDB+, EXPLORE/OPENNEXUS-E, EXPLORE/FRAME)}
\item{index}{Sequential identifier for each entry in the dataset. (Examples: 3, 1, 6)}
\item{eth_domain_institution}{Identifier for the institution within the ETH domain associated with the research project. (Examples: WSL, EPFL, PSI)}
\item{function_title}{Job title or position held by an individual associated with the research project. (Examples: Postdoctoral researcher, Deputy Director WASH R&D Center, Scientific collaborator)}
\item{first_name}{Given name of an individual associated with the research project. (Examples: Peter, Zhanyun, Mati Ullah)}
\item{surname}{Family name of an individual associated with the research project. (Examples: Schaffner, De Moraes, Feulner)}
\item{postcode}{Postal code associated with the address of an individual or institution. (Range: 0 to 9014)}
}
}
\usage{
application_metadata_applicants
}
\description{
ETH Open Research Data dataset, A colletion of metadata and report data of a CHF 15 million ETH domain program dedicated to increase the visibility and the impact of their research within the scientific community, the economy and society as a whole.
}
\keyword{datasets}
