% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/application_metadata_keywords.R
\docType{data}
\name{application_metadata_keywords}
\alias{application_metadata_keywords}
\title{application_metadata_keywords: ETH Board Open Research Data (ORD) Program Project Metadata and Report Data}
\format{
A data frame with 303 rows and 3 variables:
\describe{
\item{project_id}{Identifier for each research project within the ETH Open Research Data dataset. (Examples: EXPLORE/FRAME, EXPLORE/FAIRQUAL, CONTRIBUTE/TREENETGAPS)}
\item{index}{Sequential index number assigned to each entry in the dataset. (Range: 0 to 10)}
\item{keyword}{Descriptive tags or terms associated with the research project to indicate its focus or subject matter. (Examples: Data Processing & Visualization, predictive models, Data Science)}
}
}
\usage{
application_metadata_keywords
}
\description{
ETH Open Research Data dataset, A colletion of metadata and report data of a CHF 15 million ETH domain program dedicated to increase the visibility and the impact of their research within the scientific community, the economy and society as a whole.
}
\keyword{datasets}
