% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/details.R
\docType{data}
\name{details}
\alias{details}
\title{details: Open Research Data (ORD) Program Project Metadata and Report Data}
\format{
A tibble with 9 rows and 20 variables
\describe{
\item{project_id}{A unique identifier assigned to each project.}
\item{call_category}{The category of the project call, such as Explore or Contribute.}
\item{applicant_id}{A unique identifier assigned to the applicant.}
\item{main_applicant_institution_name}{The name of the main institution associated with the applicant.}
\item{applicant_type}{The type of applicant, such as Primary or Secondary.}
\item{applicant_title}{The title of the applicant, such as Prof. Dr.}
\item{applicant_first_name}{The first name of the applicant.}
\item{applicant_last_name}{The last name of the applicant.}
\item{applicant_gender}{The gender of the applicant, categorized as Female or Male.}
\item{applicant_institution}{The institution associated with the applicant.}
\item{applicant_department_name}{The name of the department associated with the applicant.}
\item{applicant_lab_name}{The name of the lab associated with the applicant.}
\item{applicant_orcid_id}{The ORCID identifier of the applicant.}
\item{title}{The title of the project.}
\item{acronym}{A shortened form of the project title.}
\item{abstract}{A brief summary of the project.}
\item{keywords}{Relevant keywords associated with the project.}
\item{project_start_dd_mm_yyyy}{The start date of the project in the format DD.MM.YYYY.}
\item{project_duration_months}{The duration of the project in months.}
\item{funding_requested}{The amount of funding requested for the project.}
}
}
\usage{
details
}
\description{
This data package contains metadata and report data from over 60 research projects funded by the joint Open Research Data (ORD) program of ETH Zurich, EPFL, and the four research institutes of the ETH Domain from 2020-2023. The package includes project-level metadata, such as project titles, descriptions, and categorizations (Establish, Explore, Contribute), as well as data extracted from reports submitted by each project, providing insights into the projects' objectives, methods, outcomes, and impacts.
}
\keyword{datasets}
