% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ethord-package.R
\docType{package}
\name{ethord-package}
\alias{ethord}
\alias{ethord-package}
\title{ethord: ETH Board Open Research Data (ORD) Program Project Metadata and Report Data}
\description{
ETH Open Research Data dataset, A colletion of metadata and report data of a CHF 15 million ETH domain program dedicated to increase the visibility and the impact of their research within the scientific community, the economy and society as a whole.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Global-Health-Engineering/ethord}
  \item Report bugs at \url{https://github.com/Global-Health-Engineering/ethord/issues}
}

}
\author{
\strong{Maintainer}: Lars Schöbitz \email{lschoebitz@ethz.ch} (\href{https://orcid.org/0000-0003-2196-5015}{ORCID})

Authors:
\itemize{
  \item Nicolò Massari \email{nmassari@ethz.ch} (\href{https://orcid.org/0009-0006-8421-930X}{ORCID})
  \item Elizabeth Tilley \email{tilleye@ethz.ch} (\href{https://orcid.org/0000-0002-2095-9724}{ORCID}) [funder]
}

}
\keyword{internal}
