% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ord_portal.R
\docType{data}
\name{ord_portal}
\alias{ord_portal}
\title{ord_portal: Open Research Data (ORD) Program Project Metadata and Report Data}
\format{
A tibble with 96 rows and 8 variables
\describe{
\item{url}{The URL associated with the research project, likely linking to more detailed information or resources.}
\item{title}{The title of the research project, providing a brief summary of its focus or objective.}
\item{category}{The category or classification of the research project, with possible values including Contribute, Explore, and Establish.}
\item{institutions}{The institutions or organizations involved in or supporting the research project, such as EPFL, ETH Zurich, or Empa.}
\item{data_type}{The type or classification of data associated with the research project, such as microstructure data, environmental data, or medical data.}
\item{field}{The broader field of study or domain that the research project belongs to, including Materials Science, Earth sciences, and Life sciences.}
\item{researchers}{The names of researchers or individuals involved in the project, listed in a format such as 'Last Name, First Name'.}
\item{abstract}{A brief summary or abstract of the research project, providing an overview of its objectives, methods, or findings.}
}
}
\usage{
ord_portal
}
\description{
This data package contains metadata and report data from over 60 research projects funded by the joint Open Research Data (ORD) program of ETH Zurich, EPFL, and the four research institutes of the ETH Domain from 2020-2023. The package includes project-level metadata, such as project titles, descriptions, and categorizations (Establish, Explore, Contribute), as well as data extracted from reports submitted by each project, providing insights into the projects' objectives, methods, outcomes, and impacts.
}
\keyword{datasets}
