% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposals.R
\docType{data}
\name{proposals}
\alias{proposals}
\title{proposals: Open Research Data (ORD) Program Project Metadata and Report Data}
\format{
A tibble with 25 rows and 12 variables
\describe{
\item{project_id}{A unique identifier for each project in the dataset.}
\item{cost_number}{A numerical identifier for a specific cost or expenditure within a project.}
\item{cost_personnel_senior_staff_fr}{The cost of personnel for senior staff in Swiss Francs (CHF). This is a continuous numerical measurement.}
\item{cost_personnel_postdocs_fr}{The cost of personnel for postdoctoral researchers in Swiss Francs (CHF). This is a continuous numerical measurement.}
\item{cost_personnel_other_fr}{The cost of personnel for other staff members in Swiss Francs (CHF). This is a continuous numerical measurement.}
\item{cost_personnel_students_fr}{The cost of personnel for students in Swiss Francs (CHF). This is a continuous numerical measurement.}
\item{cost_travel_fr}{The cost of travel expenses in Swiss Francs (CHF). This is a continuous numerical measurement.}
\item{cost_equipment_fr}{The cost of equipment purchases in Swiss Francs (CHF). This is a continuous numerical measurement.}
\item{cost_publication_fr}{The cost of publication expenses in Swiss Francs (CHF). This is a continuous numerical measurement.}
\item{cost_social_fr}{The cost of social expenses in Swiss Francs (CHF). This is a continuous numerical measurement.}
\item{cost_other_fr}{The cost of other miscellaneous expenses in Swiss Francs (CHF). This is a continuous numerical measurement.}
\item{cost_subcontracting_fr}{The cost of subcontracting services in Swiss Francs (CHF). This is a continuous numerical measurement.}
}
}
\usage{
proposals
}
\description{
This data package contains metadata and report data from over 60 research projects funded by the joint Open Research Data (ORD) program of ETH Zurich, EPFL, and the four research institutes of the ETH Domain from 2020-2023. The package includes project-level metadata, such as project titles, descriptions, and categorizations (Establish, Explore, Contribute), as well as data extracted from reports submitted by each project, providing insights into the projects' objectives, methods, outcomes, and impacts.
}
\keyword{datasets}
