% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_metadata.R
\docType{data}
\name{report_metadata}
\alias{report_metadata}
\title{report_metadata: ETH Board Open Research Data (ORD) Program Project Metadata and Report Data}
\format{
A data frame with 63 rows and 10 variables:
\describe{
\item{project_id}{Unique identifier for each research project within the ETH Open Research Data dataset. (Examples: CONTRIBUTE/TIMERESHDRMX, ESTABLISH/PREMISE, CONTRIBUTE/GPM-API)}
\item{project_title}{Title of the research project that outlines its objectives and scope. (Examples: Explore AiiDA for Regional Inverse Modelling of Greenhouse Gases, Building Open-Source Tools for reproducible interaction with biological ORD databases, XYT : from raw GPS data to Mobility analysis)}
\item{explore_call_number}{Identifier for the specific call under which the research project was submitted, represented as a numeric value. (Examples: 6, 5, 0)}
\item{project_starting_and_end_date}{Start and end dates of the research project formatted as YYYY-MM-DD. (Examples: 2023-09-01/2025-02-28, 2023-05-01/2025-06-30, 2023-03-01/2024-08-31)}
\item{main_applicant_responsible_for_project}{Name of the main applicant responsible for overseeing the research project. (Examples: Marco Hutter, Esther Frei, Serina Robinson)}
\item{function}{Job title or role of the individual within the research project. (Examples: Professor, Senior Scientist / Teamleader, Photon Science Division)}
\item{laboratory_department}{Department or unit within the institution where the research project is conducted. (Examples: RU Forest Resources and Management, Laboratory for Waste Management, Center for Imaging)}
\item{institution_of_main_applicant}{Name of the institution affiliated with the main applicant of the research project. (Examples: Geological Institute, Paul Scherrer Institute PSI, Paul Scherrer Institute)}
\item{report}{Classification of the type of report associated with the research project. (Examples: Final, other, Intermediate)}
\item{reporting_period}{Timeframe during which project activities are reported, formatted as start date/end date. (Examples: 2023-05-01/2024-07-31, 2025-04-01/2025-03-31, 2022-10-01/2023-09-30)}
}
}
\usage{
report_metadata
}
\description{
ETH Open Research Data dataset, A colletion of metadata and report data of a CHF 15 million ETH domain program dedicated to increase the visibility and the impact of their research within the scientific community, the economy and society as a whole.
}
\keyword{datasets}
